Fix test failures with authlib 1.6.2
The test failures are probably caused by this change in authlib:
https://github.com/authlib/authlib/commit/a4a9792bf371c7b1d2738bd0c9edda76b66eed41
in our tests, `code` was present in `kwargs` when checked but its
value was `None` (a default value set along the way); with the
previous authlib code this was incorrectly allowed, but now it is
correctly rejected.

Somebody more familiar with OAuth should check this, but I *think*
it's correct to just fake the value. Looking at
https://docs.authlib.org/en/latest/client/oauth2.html , it seems
the code is supposed to be provided by the authorization server.
Once we've generated an authorization URL we're supposed to visit
it and "grant the authorization", and it should
"redirect you back to your site with a code and state arguments".
If I'm following things correctly, we're just faking this step
in the tests - we don't really have an authorization server,
we're just making up a response as if it came from one. So it
should be correct to include a dummy `code` value in the made-up
responses.

Signed-off-by: Adam Williamson <awilliam@redhat.com>