Fix test_get_overrides_rss for builds with +
In recent runs, the data dump used by the CI tests happened to
have a build with a + in its name in the buildroot overrides.
This causes a problem because pyramid (which constructs the
string used as `text` in the RSS entries) does not quote +
characters, but `urllib.parse.quote` (which we use to construct
the expected results to compare against) does quote them. As
suggested by @mattiaverga , let's fix this by adding + to the
'safe' characters for `quote` so it won't quote it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>