resultsdb consumer: skip status update appropriately for waiting
We can skip updating the gating status if the current status is
'waiting', and the message we're handling is for a QUEUED or
RUNNING result. This should substantially reduce backups in the
consumer (and load on greenwave) by making us not update the
status for every single QUEUED or RUNNING message emitted for a
newly-created update (which can be hundreds), when we already
set the gating status for newly-created updates to 'waiting'.

Fixes: #6000
Signed-off-by: Adam Williamson <awilliam@redhat.com>