Make client `updates download` use koji --fallback-unsigned
When we're trying to download signed packages, use the
--fallback-unsigned option to tell Koji to fall back to getting
unsigned packages if it can't get signed ones. This is just what
we used to do anyway, and better than bailing out or downloading
nothing at all.

This was always planned as a follow-up to
https://github.com/fedora-infra/bodhi/pull/5859 , but there was
a delay in getting a Koji release with the feature. I've now
backported the Koji feature to Fedora 42+ and EPEL 10+, so it
should be reasonably safe to land this and send an update to the
same branches.

Signed-off-by: Adam Williamson <awilliam@redhat.com>