Update automatically-created update test gating status on creation
https://github.com/fedora-infra/bodhi/pull/3519 cleaned up the
automatic update creation process a lot, but I think introduced a
minor bug. It removed the setting of the update gating status at
automatic creation time. Instead, the gating status would only be
set to something other than `None` when the update was marked as
'testing' (or by any other process which happens to set it in the
mean time).

We noticed this in
https://github.com/fedora-infra/bodhi/issues/6000#issuecomment-3677952903
At first I was going to just drop the gating status update from
SignedHandler , but on closer inspection I think it's actually
correct to move it back to AutomaticUpdateHandler so we get an
initial gating status immediately upon update creation. Right now
it *usually* happens to get updated very quickly anyway -
probably in response to a QUEUED test result being created, in
most cases - but sometimes it doesn't, e.g. in
https://bodhi.fedoraproject.org/updates/FEDORA-2025-73c49c68e3
the initial gating status was set 36 minutes after the update
was created.

We do drop the initial set to waiting before calling
update_test_gating_status, because it's pointless and just causes
the weird status churn that @mattiaverga noticed.

Signed-off-by: Adam Williamson <awilliam@redhat.com>